+++ neurocaps/_utils/_check_kwargs.py
@@ -5,7 +5,7 @@ def _check_kwargs(defaults, **kwargs):
 
     if kwargs:
         invalid_kwargs = {k: v for k, v in kwargs.items() if k not in plot_dict}
-        if len(invalid_kwargs) > 0:
+        if invalid_kwargs:
             print(f"Invalid kwargs arguments used and will be ignored {invalid_kwargs}.")
 
     return plot_dict